+++ src/main/java/ua/goit/controller/projects/DeleteProjectServlet.java
@@ -24,10 +24,16 @@ public class DeleteProjectServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String name = req.getParameter("name");
-        ProjectDAO projectDAOForDelete = projectsRepository.findByUniqueParameter("projectName", name).orElseThrow();
-        projectsRepository.deleteByParameter("projectName", name);
-        req.setAttribute("name", projectDAOForDelete.getProjectName());
-        req.getRequestDispatcher("/view/projects/deleteProject.jsp").forward(req, resp);
+        if (projectsRepository.findByUniqueParameter("projectName", name).isEmpty()) {
+            ServletException ex = new ServletException(String.format("The project with %s does not exist", name));
+            req.setAttribute("message", ex.getMessage());
+            req.getRequestDispatcher("/view/errorPage.jsp").forward(req, resp);
+        } else {
+            ProjectDAO projectDAOForDelete = projectsRepository.findByUniqueParameter("projectName", name).orElseThrow();
+            projectsRepository.deleteByParameter("projectName", name);
+            req.setAttribute("name", projectDAOForDelete.getProjectName());
+            req.getRequestDispatcher("/view/projects/deleteProject.jsp").forward(req, resp);
+        }
     }
 
 }
